Add logic to verify every wiring call
This change adds logic to check that every wiring call is valid. For
example, this check catches the case where we try to wire things to a
definition after "EndCircuit" has been called, or if we try to use an
instance that was not placed inside a defn.

TODO(rsetaluri): Add a gate to make this only happen in a certain mode
(like debug mode), since there may be performance implications of this
call happening for every wire.